semantic-kernel-demo/requirements.txt
semantic-kernel==1.28.0
python-dotenv==1.0.0
httpx==0.27.0
